add custom action modification statement
This allows to reference actions groups that have no representations in
an MPS model by specifying the name. This does not feature any static
checks inside of MPS. If in correct data is supplied it might NPE at
runtime. The action group name can be found in the xml declaration inside
of MPS for instance /workbench/resources/source/idea/PlatformActions.xml