TopDownLayout: IndentLayout: Line is wrapped depending on the structure of the editor
There are preferred positions for line breaks. In an expression
containing a function call, it is preferred to insert a line break
before or after the function call and not inside of it.
If the function call has a very long parameter list, it might be
preferred to have the line break after one the parameter instead of
having lots of empty space on a line.

The method for rating a line break position still needs some work.