TopDownLayout,Diagram: Cells can now grow and push in x and y direction
A cell inside a grid based layout (horizontal, vertical, vertical grid)
can fill the column/row size if it has the grow attribute set (A
column/row is set to the size of the biggest cell. The other cells can
then be only aligned, or they can grow to the size of the column/row).
The push attribute also makes the parent grow.

This is required for subdiagrams. They have to use up all the remaining
space inside a box.