TopDownLayout: overflow-x and overflow-y style attributes
If you have a large cell, that is wider than the right margin, this cell
would stretch the whole editor and cause all cells to wrap after the
right margin.

In this case, you can set overflow-x and the layouter will layout the
surrounding cells as if the width is 0.

Make sure, that there are no other cells in the same line. Otherwise,
they will overlap.

This is used for listings in the doc language. They do not wrap and if
it contains long lines, the whole document was as wide as the listing.